Dbg ITM: Prevent 'bx in trigger_out
The itm_trace_qualificator expects 0 as trace data if no valid data
is available (wb_freeze = 1). This expectation was violated when
integrating the clock-domain crossings. This in turn resulted in
trigger_out being 'bx if no valid data was available, which might
result in a simulation mismatch on the FPGA.