SRAM: Updates and integrate
Note that this SRAM implementation requires new parameters to select
the RAM implementation (plain, block ram, etc.).

YOU WILL NEED TO ADJUST YOUR DESIGNS!

Plain Memory: All memory implementations are expected to have one
cycle delay between setting the address and the output of the
data. This was not the case in the current implementation leading to
memory errors in burst mode.

Fix minor bugs and style issues

 * Add MEM_FILE parameter to toplevel

 * Fix calculation of # of blocks in Spartan 6 BRAM

 * Coding style/indentation/whitespace fixes

 * Various style fixes (no functional changes)